+++ NEWS.md
@@ -43,4 +43,9 @@
 ### Version 0.0.0.8 (2026-09-18)
 * The parallel processing issue may be related to how rasters are handled in the terra package and its interaction with parallel processing. 
 * As such, raster loading is added inside the parallel processing loop. 
+* This fixed the issues. Removing additional text to diagnose issues and updating parallel processing documentation to state that it is now possible for up to 8 cores
+* Added a file moving code to classifyHMD to remove cleaned files to classified folder
+* Moved classifyHMD's internal functions outside of classifyHMD. This may break the function. Will need to check when needing to run again. 
+
+